+++ src/components/layouts/icons/index.ts
@@ -12,6 +12,8 @@ import { InformationCircleIcon } from './InformationCircleIcon'
 import { ChevronDownIcon, ChevronDownIconSolid } from './ChevronDownIcon'
 import { WarningTriangleIcon } from './WarningTriangleIcon'
 import { ExternalLinkIcon } from './ExternalLinkIcon'
+import { MobilePhoneIcon } from './MobilePhoneIcon'
+import { DesktopMonitorIcon } from './DesktopMonitorIcon'
 
 export {
   EyeIcon,
@@ -29,4 +31,6 @@ export {
   VerificationIconSolid,
   WarningTriangleIcon,
   ExternalLinkIcon,
+  MobilePhoneIcon,
+  DesktopMonitorIcon,
 }